requirements.txt
streamlit==1.9.0
numpy==1.23.0
Pillow==8.2.0
tensorflow==2.5.0
opencv-python==4.5.4.58